Create bowtie2_auto.sh
Ensure you change the working directory on page 35.
Also you need the 6 .bt2 references in the current directory.

For this script, all files are assumed to be coming from the cutadapt pipeline and should end in " *_trim.fastq.gz " 

The script will count the unique prefixes in the current directory, and prompt an output folder. It will then generate a temporary script that can be submitted to the slurm scheduler to process in parallel. Each script will map one trimmed fastq file against the GRCh38 reference, and output as SAM file. These are huge so we convert to BAM using samtools/1.6. Then you need to sort the BAM in samtools, and finally you can generate an index file (.bam.bai) which needs to be downloaded with the bam in order to view in IGV.